requirements.txt
langchain>=0.0.335
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.24.3
matplotlib>=3.8.2
seaborn>=0.13.0
scikit-learn>=1.3.2
gsw>=3.6.17
netCDF4>=1.6.5
xarray>=2023.11.0
cartopy>=0.22.0
jupyter>=1.0.0
notebook>=7.0.6
ipykernel>=6.27.1
plotly>=5.18.0
scipy>=1.11.4
statsmodels>=0.14.0
langchain-community>=0.0.10
markdown>=3.4.0 